Load category as tab and sample pixel of an image for color
