+++ SceneEnhancementLabeling/View/ExtractColorPage.xaml
@@ -1,0 +1,67 @@
+﻿<Page x:Class="SceneEnhancementLabeling.View.ExtractColorPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:SceneEnhancementLabeling.View"
+      mc:Ignorable="d" 
+      d:DesignHeight="300" d:DesignWidth="300"
+      Title="ExtractColorPage"
+      DataContext="{Binding ExtractColor, Source={StaticResource Locator}}">
+
+    <Page.InputBindings>
+        <KeyBinding Key="O" Modifiers="Control" Command="{Binding OpenFolderCommand}"/>
+    </Page.InputBindings>
+    <Grid>
+        <DockPanel>
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="3*"/>
+                    <ColumnDefinition/>
+                </Grid.ColumnDefinitions>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="4*"/>
+                    <RowDefinition/>
+                </Grid.RowDefinitions>
+                <Grid>
+                    <Image Source="{Binding Bitmap}" Stretch="Uniform"/>
+                    <Button HorizontalAlignment="Left" VerticalAlignment="Center" Content="&lt;" Width="30" Height="20" Visibility="{Binding CanPrevious, Converter={StaticResource BooleanToVisibilityConverter}}" Command="{Binding PreviousCommand}"/>
+                    <Button HorizontalAlignment="Right" VerticalAlignment="Center" Content="&gt;" Width="30" Height="20" Visibility="{Binding CanNext, Converter={StaticResource BooleanToVisibilityConverter}}" Command="{Binding NextCommand}"/>
+                </Grid>
+                <Grid Grid.Row="0" Grid.Column="1">
+                    <Grid.RowDefinitions>
+                        <RowDefinition/>
+                        <RowDefinition/>
+                        <RowDefinition/>
+                        <RowDefinition/>
+                        <RowDefinition/>
+                    </Grid.RowDefinitions>
+                    <Grid Margin="20">
+                        <Rectangle Fill="{Binding Color0}"/>
+                        <TextBlock Text="{Binding Color0.Color}"/>
+                    </Grid>
+                    <Grid Margin="20" Grid.Row="1">
+                        <Rectangle Fill="{Binding Color1}"/>
+                        <TextBlock Text="{Binding Color1.Color}"/>
+                    </Grid>
+                    <Grid Margin="20" Grid.Row="2">
+                        <Rectangle Fill="{Binding Color2}"/>
+                        <TextBlock Text="{Binding Color2.Color}"/>
+                    </Grid>
+                    <Grid Margin="20" Grid.Row="3">
+                        <Rectangle Fill="{Binding Color3}"/>
+                        <TextBlock Text="{Binding Color3.Color}"/>
+                    </Grid>
+                    <Grid Margin="20" Grid.Row="4">
+                        <Rectangle Fill="{Binding Color4}"/>
+                        <TextBlock Text="{Binding Color4.Color}"/>
+                    </Grid>
+                </Grid>
+                <StackPanel Orientation="Horizontal" Grid.Row="1" Grid.Column="0">
+                    <Button Content="Browse..." Width="100" Height="50" Command="{Binding OpenFolderCommand}" Margin="10"/>
+                    <Button Content="Start" Command="{Binding StartCommand}" Width="100" Height="50" Margin="10"/>
+                </StackPanel>
+            </Grid>
+        </DockPanel>
+    </Grid>
+</Page>
