+++ SceneEnhancementLabeling/View/LabelingPage.xaml
@@ -1,0 +1,84 @@
+﻿<Page x:Class="SceneEnhancementLabeling.View.LabelingPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:SceneEnhancementLabeling.View"
+      xmlns:common="clr-namespace:SceneEnhancementLabeling.Common"
+      xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+      mc:Ignorable="d" 
+      d:DesignHeight="300" d:DesignWidth="300"
+      Title="LabelingPage"
+      DataContext="{Binding Labeling, Source={StaticResource Locator}}">
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+        <TextBlock Text="Labeling" FontWeight="Bold" FontSize="30" Margin="12"/>
+        <Grid Grid.Row="1" Margin="12,0">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="2*"/>
+                <ColumnDefinition Width="*"/>
+            </Grid.ColumnDefinitions>
+            <Grid>
+                <Button Content="Browse..." Width="100" Height="30" Command="{Binding OpenFolderCommand}" Visibility="{Binding IsBrowseEnabled, Converter={StaticResource BooleanToVisibilityConverter}}"/>
+                <Grid Margin="12">
+                    <Grid>
+                    <Image Source="{Binding Bitmap}" Stretch="Uniform">
+                        <i:Interaction.Behaviors>
+                            <common:ImageBehaviorMouseDownPointSampleToColor TestSource="{Binding Source, Mode=TwoWay}" SelectedColor="{Binding SelectedColor, Mode=TwoWay}"/>
+                        </i:Interaction.Behaviors>
+                    </Image>
+                        <Image  Source="{Binding Source}" Stretch="Uniform" Opacity="0.2" IsHitTestVisible="False"/>
+                    </Grid>
+                    <Button HorizontalAlignment="Left" VerticalAlignment="Center" Content="&lt;" Width="30" Height="20" Visibility="{Binding CanPrevious, Converter={StaticResource BooleanToVisibilityConverter}}" Command="{Binding PreviousCommand}"/>
+                    <Button HorizontalAlignment="Right" VerticalAlignment="Center" Content="&gt;" Width="30" Height="20" Visibility="{Binding CanNext, Converter={StaticResource BooleanToVisibilityConverter}}" Command="{Binding NextCommand}"/>
+                </Grid>
+            </Grid>
+            <Grid Grid.Column="1">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="Auto"/>
+                    <RowDefinition/>
+                </Grid.RowDefinitions>
+                <StackPanel>
+                    <TextBlock Text="1. Color Labeling" FontSize="20" FontWeight="Bold"/>
+                    <TabControl ItemsSource="{Binding Category}" SelectedIndex="{Binding CategoryIndex, Mode=TwoWay}">
+                        <TabControl.ItemTemplate>
+                            <DataTemplate>
+                                <TextBlock Text="{Binding Name}"/>
+                            </DataTemplate>
+                        </TabControl.ItemTemplate>
+                        <TabControl.ContentTemplate>
+                            <DataTemplate>
+                                <Grid Height="200" Margin="12">
+                                    <Grid.RowDefinitions>
+                                        <RowDefinition/>
+                                        <RowDefinition/>
+                                        <RowDefinition/>
+                                    </Grid.RowDefinitions>
+                                    <StackPanel Orientation="Horizontal" VerticalAlignment="Center">
+                                        <Rectangle Fill="{Binding Color0, Mode=TwoWay}" Width="100" Height="50"/>
+                                        <TextBlock Text="{Binding Color0, Mode=TwoWay}" VerticalAlignment="Center" Margin="12"/>
+                                    </StackPanel>
+                                    <StackPanel Grid.Row="1" Orientation="Horizontal" VerticalAlignment="Center">
+                                        <Rectangle Fill="{Binding Color1, Mode=TwoWay}" Width="100" Height="50"/>
+                                        <TextBlock Text="{Binding Color1, Mode=TwoWay}" VerticalAlignment="Center" Margin="12"/>
+                                    </StackPanel>
+                                    <StackPanel Grid.Row="2" Orientation="Horizontal" VerticalAlignment="Center">
+                                        <Rectangle Fill="{Binding Color2, Mode=TwoWay}" Width="100" Height="50"/>
+                                        <TextBlock Text="{Binding Color2, Mode=TwoWay}" VerticalAlignment="Center" Margin="12"/>
+                                    </StackPanel>
+                                </Grid>
+                            </DataTemplate>
+                        </TabControl.ContentTemplate>
+                    </TabControl>
+                </StackPanel>
+                <StackPanel Grid.Row="1" Margin="0,12,0,0">
+                    <TextBlock Text="2. Component Labeling" FontSize="20" FontWeight="Bold"/>
+                </StackPanel>
+            </Grid>
+        </Grid>
+    </Grid>
+</Page>
