+++ SceneEnhancementLabeling/packages.config
@@ -3,4 +3,5 @@
   <package id="CommonServiceLocator" version="1.3" targetFramework="net452" />
   <package id="MvvmLight" version="5.2.0.0" targetFramework="net452" />
   <package id="MvvmLightLibs" version="5.2.0.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net452" />
 </packages>
